Adds initial routes and website layout
This adds the resourceful routes to Actor, Movie, and
production_company.  They each have routes for all of the CRUD
functions.

This also adds a layout for all of the views to inherit from.  The
layout mainly comes from Bootstrap's basic template.  I merged that with
the default Laravel template.