Adds first views for Movie routes
I created the files for the different views that will be needed (create,
edit, index, and show).  I have completed index and show.  I also wired
the controller to return the correct view with the correct data.

Finally, I set the LC_MONETARY locale in PHP so that the money will show
up as USD.