Adds first views for Actor
This adds the stubs for all of the CRUD views but so far the only that
work are the index and show.  This also wires up the ActorController to
pass the correct data to the views.

I also added a new function to the Role model to more easily calculate
an actor's salary.