Adds create, update, and destroy to controllers
This is a rather large change but it gets create, update, and destroy
working on the `Actor`, `Movie`, and `production_company` controllers.
It uses the Laravel Collective's HTML/Form library to create the forms
for all of the views.  I found that to be a quick and easy way to
generate them.  The destroy has been implemented as a small form on the
index pages with just a single delete button.  That was the only way I
could figure out how to send a 'DELETE' request to the server.