Adds deletion capability to views
In order to send a DELETE request, I had to have forms so I created some
simple forms that only have a single button to delete on them and I
updated all of the index views to use this.

I also fixed a few spots where the form wasn't filling correctly when a
model already exists.  I updated the default value to `null` and that
corrected the error.