Adds resources for Scripts
This adds a model and factory for scripts.  It also updates the
`movies.show` view to include the Script stats.  The script stats are
the number of lines spoken per actor in a movie, the number of words
spoken by an actor in a film, and the number of references made to a
character in a film.